requirements.txt
anyio==3.4.0
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
dnspython==2.1.0
email-validator==1.1.3
fastapi==0.70.1
h11==0.12.0
httptools==0.2.0
idna==3.3
importlib-metadata==4.9.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
orjson==3.6.5
pydantic==1.8.2
PyJWT==2.3.0
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
sniffio==1.2.0
starlette==0.16.0
twilio==7.4.0
typing_extensions==4.0.1
ujson==4.3.0
urllib3==1.26.7
uvicorn==0.15.0
watchgod==0.7
websockets==10.1
zipp==3.6.0
